ai_assistant/requirements.txt
# Core
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# LLM providers (install at least one)
google-genai>=1.0.0
openai>=1.12.0

# Utilities
python-dotenv>=1.0.0
